catcoder/pyproject.toml
[project]
name = "catcoder"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "backoff==2.2.1",
    "chromadb==0.5.5",
    "datasets==2.14.6",
    "jinja2==3.1.2",
    "langchain==0.2.14",
    "langchain-community==0.2.12",
    "langchain-core==0.2.33",
    "numpy==1.26.1",
    "openai>=1.30.0",
    "pyarrow<21",
    "python-dotenv==1.0.0",
    "rank-bm25==0.2.2",
    "sentence-transformers",
    "torch",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
